+++ libs/pkcs11/src/CSP/VerificaConCIE.cpp
@@ -20,7 +20,7 @@ extern "C" {
 CK_RV CK_ENTRY verificaConCIE(const char* inFilePath, const char* proxyAddress,
                               int proxyPort, const char* usrPass);
 CK_RV CK_ENTRY getNumberOfSign(void);
-CK_RV CK_ENTRY getVerifyInfo(int index, struct verifyInfo_t* vInfos);
+CK_RV CK_ENTRY getVerifyInfo(u_int64_t index, struct verifyInfo_t* vInfos);
 CK_RV CK_ENTRY estraiP7m(const char* inFilePath, const char* outFilePath);
 }
